refactor of utils
