+++ crate/src/position_utils/get_lazy_statics.rs
@@ -1,0 +1,142 @@
+use crate::constants::MATH_PI;
+use crate::id_generator::IdGenerator;
+use std::collections::HashMap;
+use super::basic_utils::{Point,Line,BasicUtils};
+
+pub static OBSTACLES_LENGTH: [usize; 2] = [4, 5];
+
+pub static RAW_POINTS: [(f32, f32); 9] = [
+  (600.0, 100.0),
+  (900.0, 100.0),
+  (900.0, 300.0),
+  (600.0, 300.0),
+  // end here
+  (700.0, 400.0),
+  (900.0, 400.0),
+  (900.0, 600.0),
+  (700.0, 600.0),
+  (600.0, 500.0),
+  // end here
+];
+
+pub struct GetLazyStatics {}
+
+impl GetLazyStatics {
+  fn insert_lines_to_graph<'a>(
+    ref_graph: &mut HashMap<u32, Vec<&'a Point>>,
+    point_a: &'a Point,
+    point_b: &'a Point
+   ) {
+    match ref_graph.get_mut(&point_a.id) {
+      Some(connected_points_list) => {
+        connected_points_list.push(point_b);
+      }
+      None => {
+        ref_graph.insert(point_a.id, vec![point_b]);
+      }
+    };
+  }
+
+  pub fn get_obstacles_points() -> Vec<Point>{
+    RAW_POINTS
+      .iter()
+      .map(|(x, y)| Point {
+        id: IdGenerator::generate_id() as u32,
+        x: *x,
+        y: *y,
+      })
+      .collect()
+  }
+
+  pub fn get_obstacles_lines(points: &'static Vec<Point>) -> Vec<Line<'static>> {
+    let mut obstacle_index = 0;
+    let mut obstacle_start_point_index = 0;
+
+    points
+      .iter()
+      .enumerate()
+      .map(|(index, point)| {
+        let obstacle_last_point_index = obstacle_start_point_index + OBSTACLES_LENGTH[obstacle_index] - 1;
+        let connected_point_index =
+        if index == obstacle_last_point_index {
+          let prev_obstacle_start_point_index = obstacle_start_point_index;
+          obstacle_start_point_index += OBSTACLES_LENGTH[obstacle_index];
+          obstacle_index += 1;
+          prev_obstacle_start_point_index
+        } else {
+          index + 1
+        };
+
+        Line {
+          p1: point,
+          p2: &points[connected_point_index],
+        }
+      })
+      .collect()
+  }
+
+  pub fn get_permanent_connection_graph(
+    obstacles_points: &'static Vec<Point>,
+    obstacles_lines: &'static Vec<Line>,
+  ) -> HashMap<u32, Vec<&'static Point>> {
+    let mut graph: HashMap<u32, Vec<&Point>> = HashMap::new();
+
+    obstacles_lines.iter().for_each(|line| {
+      GetLazyStatics::insert_lines_to_graph(&mut graph, &line.p1, &line.p2);
+      GetLazyStatics::insert_lines_to_graph(&mut graph, &line.p2, &line.p1);
+    });
+
+    let mut obstacle_index = 0;
+    let mut obstacle_start_point_index = 0;
+    let mut grouped_obstacles_points: Vec<Vec<&Point>> = vec![vec![]];
+
+    obstacles_points
+      .iter()
+      .enumerate()
+      .for_each(|(index, point)| {
+        if index == obstacle_start_point_index + OBSTACLES_LENGTH[obstacle_index] - 1 {
+          grouped_obstacles_points[obstacle_index].push(point);
+          grouped_obstacles_points.push(vec![]);
+          obstacle_start_point_index += OBSTACLES_LENGTH[obstacle_index];
+          obstacle_index += 1;
+        } else {
+          grouped_obstacles_points[obstacle_index].push(point);
+        };
+      });
+
+      grouped_obstacles_points.iter().enumerate().for_each(|(index_a, obstacle_a)| {
+        for index_b in index_a + 1..grouped_obstacles_points.len() {
+
+          let obstacle_b = &grouped_obstacles_points[index_b];
+          obstacle_a.iter().for_each(|point_a| {
+            obstacle_b.iter().for_each(|point_b| {
+
+              let new_line = Line {
+                p1: point_a,
+                p2: point_b,
+              };
+              let mut is_intersect = false;
+              obstacles_lines.iter().for_each(|obstacle_line| {
+                if obstacle_line.p1.id != point_a.id
+                  && obstacle_line.p1.id != point_b.id
+                  && obstacle_line.p2.id != point_a.id
+                  && obstacle_line.p2.id != point_b.id
+                  && BasicUtils::check_intersection(&new_line, obstacle_line)
+                {
+                  is_intersect = true;
+                };
+              });
+
+              if !is_intersect {
+                graph.get_mut(&point_a.id).unwrap().push(&point_b);
+                graph.get_mut(&point_b.id).unwrap().push(&point_a);
+              }
+            })
+          });
+
+        }
+      });
+
+    graph
+  }
+}
